requirements.txt
playwright==1.49.0
pandas==3.0.0
openpyxl==3.1.5
python-dotenv==1.2.1
